cli/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "prescale-cli"
version = "0.1.0"
description = "CLI for Prescale - Predictive Infrastructure Intelligence Platform"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    { name = "Prescale Team" }
]
keywords = ["kubernetes", "ml", "infrastructure", "cli", "devops"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "click>=8.0.0",
    "httpx>=0.24.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
prescale = "prescale_cli.main:main"

[project.urls]
Homepage = "https://github.com/pyjeebz/prescale"
Documentation = "https://github.com/pyjeebz/prescale#readme"
Repository = "https://github.com/pyjeebz/prescale"

[tool.hatch.build.targets.wheel]
packages = ["src/prescale_cli"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/README.md",
]
